fix merge error and compiler warning
FossilOrigin-Name: 5c0afe70a5ee2b378896bf30426823dc5ae4e95c

+++ src/pragma.c
@@ -412,7 +412,7 @@ void sqlite3Pragma(
       newMax = atoi(zRight);
     }
     if( ALWAYS(pBt) ){
-      newMax = sqlite3BtreeMaxPageCount(pBt, newMax);
+      newMax = sqlite3BtreeMaxPageCount(pBt, (int)newMax);
     }
     returnSingleInt(pParse, "max_page_count", &newMax);
   }else
